+++ packages/dev/src/main.ts
@@ -71,6 +71,15 @@ export interface Features {
    */
   images?: {
     enabled?: boolean
+
+    /**
+     * Allowed URL patterns for remote images, described as an array of regular
+     * expression strings. This list will be merged with  the `remote_images`
+     * configuration property.
+     *
+     * {@link} https://docs.netlify.com/image-cdn/overview/#remote-path
+     */
+    remoteURLPatterns?: string[]
   }
 
   /**
@@ -152,6 +161,7 @@ export class NetlifyDev {
     static: boolean
   }
   #headersHandler?: HeadersHandler
+  #imageRemoteURLPatterns: string[]
   #imageHandler?: ImageHandler
   #logger: Logger
   #projectRoot: string
@@ -184,6 +194,7 @@ export class NetlifyDev {
       static: options.staticFiles?.enabled !== false,
     }
     this.#functionsServePath = path.join(projectRoot, '.netlify', 'functions-serve')
+    this.#imageRemoteURLPatterns = options.images?.remoteURLPatterns ?? []
     this.#logger = options.logger ?? globalThis.console
     this.#serverAddress = options.serverAddress
     this.#projectRoot = projectRoot
@@ -453,27 +464,32 @@ export class NetlifyDev {
     }
 
     if (this.#features.edgeFunctions) {
-      const env = Object.entries(envVariables).reduce<Record<string, string>>((acc, [key, variable]) => {
-        if (
-          variable.usedSource === 'account' ||
-          variable.usedSource === 'addons' ||
-          variable.usedSource === 'internal' ||
-          variable.usedSource === 'ui' ||
-          variable.usedSource.startsWith('.env')
-        ) {
-          return {
+      const edgeFunctionsEnv = {
+        // User-defined env vars + documented runtime env vars
+        ...Object.entries(envVariables).reduce<Record<string, string>>(
+          (acc, [key, variable]) => ({
             ...acc,
             [key]: variable.value,
-          }
-        }
-
-        return acc
-      }, {})
+          }),
+          {},
+        ),
+        // Add runtime env vars that we've set ourselves so far. These are "internal" env vars,
+        // part of the runtime emulation. They've already been populated on this process's env, which
+        // is needed to make other dev features work. These are different than the "documented" runtime
+        // env vars, in that they are implementation details, needed to make our features work.
+        ...Object.keys(runtime.envSnapshot).reduce<Record<string, string>>(
+          (acc, key) => ({
+            ...acc,
+            [key]: runtime.env.get(key) ?? '',
+          }),
+          {},
+        ),
+      }
 
       const edgeFunctionsHandler = new EdgeFunctionsHandler({
         configDeclarations: this.#config?.config.edge_functions ?? [],
         directories: [this.#config?.config.build.edge_functions].filter(Boolean) as string[],
-        env,
+        env: edgeFunctionsEnv,
         geolocation: mockLocation,
         logger: this.#logger,
         siteID,
@@ -532,8 +548,12 @@ export class NetlifyDev {
     }
 
     if (this.#features.images) {
+      const remoteImages = [...(this.#config?.config?.images?.remote_images ?? []), ...this.#imageRemoteURLPatterns]
+
       this.#imageHandler = new ImageHandler({
-        imagesConfig: this.#config?.config.images,
+        imagesConfig: {
+          remote_images: remoteImages,
+        },
         logger: this.#logger,
       })
     }