+++ packages/dev-utils/CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## [3.3.0](https://github.com/netlify/primitives/compare/dev-utils-v3.2.2...dev-utils-v3.3.0) (2025-07-15)
+
+
+### Features
+
+* add `@netlify/nuxt` Nuxt module ([#344](https://github.com/netlify/primitives/issues/344)) ([efdf8df](https://github.com/netlify/primitives/commit/efdf8dfb22ffe33b557d2140fd3a3152e23cc95c))
+
+## [3.2.2](https://github.com/netlify/primitives/compare/dev-utils-v3.2.1...dev-utils-v3.2.2) (2025-07-04)
+
+
+### Bug Fixes
+
+* **deps:** move tmp-promise to prod deps ([#329](https://github.com/netlify/primitives/issues/329)) ([f8d46bf](https://github.com/netlify/primitives/commit/f8d46bf759e15ed866436838610ddadbb06c0de1))
+
 ## [3.2.1](https://github.com/netlify/primitives/compare/dev-utils-v3.2.0...dev-utils-v3.2.1) (2025-06-18)